refactored queries for swap orders
