+++ packages/yoroi-extension/app/containers/swap/orders/OrdersPage.js
@@ -6,26 +6,25 @@ import Table from '../../../components/common/table/Table';
 import CancelSwapOrderDialog from '../../../components/swap/CancelOrderDialog';
 import AssetPair from '../../../components/common/assets/AssetPair';
 import Tabs from '../../../components/common/tabs/Tabs';
-import { useRichCompletedOrders, useRichOpenOrders, useRichOrders } from '../hooks';
+import type { MappedOrder } from './hooks';
+import { useRichOrders } from './hooks';
 import type { StoresAndActionsProps } from '../../../types/injectedProps.types';
 import { SwapPoolLabel } from '../../../components/swap/SwapPoolComponents';
 import ExplorableHashContainer from '../../widgets/ExplorableHashContainer';
 import { truncateAddressShort } from '../../../utils/formatters';
 import { Quantities } from '../../../utils/quantities';
-import { PRICE_PRECISION } from '../../../components/swap/common';
 import { fail, forceNonNull, maybe, noop } from '../../../coreUtils';
 import type { RemoteTokenInfo } from '../../../api/ada/lib/state-fetch/types';
 import { useSwap } from '@yoroi/swap';
 import { addressBech32ToHex } from '../../../api/ada/lib/cardanoCrypto/utils';
-import {
-  getTransactionFeeFromCbor,
-  getTransactionTotalOutputFromCbor,
-} from '../../../api/ada/transactions/utils';
+import { getTransactionFeeFromCbor, getTransactionTotalOutputFromCbor, } from '../../../api/ada/transactions/utils';
 import { SelectedExplorer } from '../../../domain/SelectedExplorer';
 import type { CardanoConnectorSignRequest } from '../../../connector/types';
 import { genLookupOrFail } from '../../../stores/stateless/tokenHelpers';
 import moment from 'moment';
 import { signTransactionHex } from '../../../api/ada/transactions/signTransactionHex';
+import { createFormattedTokenValues } from './util';
+import type { FormattedTokenValue } from './util';
 
 type ColumnContext = {|
   completedOrders: boolean,
@@ -83,116 +82,6 @@ const orderColumns: Array<Column> = [
   },
 ];
 
-export type FormattedTokenValue = {|
-  value: string,
-  formattedValue: string,
-  ticker: string,
-|};
-
-function createFormattedTokenValues({
-  entries,
-  order,
-  defaultTokenInfo,
-}: {|
-  entries: Array<{| id: string, amount: string |}>,
-  order: any,
-  defaultTokenInfo: RemoteTokenInfo,
-|}): Array<FormattedTokenValue> {
-  const tokenAmountMap = entries.reduce(
-    (map, v) => ({ ...map, [v.id]: Quantities.sum([map[v.id] ?? '0', v.amount]) }),
-    {}
-  );
-  const ptDecimals = forceNonNull(defaultTokenInfo.decimals);
-  // $FlowIgnore[prop-missing]
-  const defaultTokenValue = tokenAmountMap[''] ?? tokenAmountMap['.'] ?? '0';
-  const formattedTokenValues = [
-    {
-      value: defaultTokenValue,
-      formattedValue: Quantities.format(defaultTokenValue, ptDecimals, ptDecimals),
-      ticker: defaultTokenInfo.ticker ?? '-',
-    },
-  ];
-  [order.from.token, order.to.token].forEach(t => {
-    if (t.id !== '' && t.id !== '.') {
-      maybe(tokenAmountMap[t.id], v => {
-        const formattedValue = Quantities.format(v, t.decimals, t.decimals);
-        formattedTokenValues.push({
-          value: v,
-          formattedValue,
-          ticker: t.ticker ?? '-',
-        });
-      });
-    }
-  });
-  return formattedTokenValues;
-}
-
-function mapOrderAssets(
-  order: any,
-  defaultTokenInfo: RemoteTokenInfo
-): {|
-  price: string,
-  amount: string,
-  totalValues: ?Array<FormattedTokenValue>,
-  from: any,
-  to: any,
-|} {
-  const price = Quantities.quotient(order.from.quantity, order.to.quantity);
-  const fromDecimals = order.from.token?.decimals ?? 0;
-  const toDecimals = order.to.token?.decimals ?? 0;
-  const priceDenomination = fromDecimals - toDecimals;
-  const formattedPrice = Quantities.format(price, priceDenomination, PRICE_PRECISION);
-  const formattedToQuantity = Quantities.format(
-    order.to.quantity,
-    toDecimals,
-    toDecimals
-  );
-  const formattedAttachedValues = maybe(order.valueAttached, val =>
-    createFormattedTokenValues({
-      entries: val.map(({ token: id, amount }) => ({ id, amount })),
-      order,
-      defaultTokenInfo,
-    })
-  );
-  return {
-    price: formattedPrice,
-    amount: formattedToQuantity,
-    totalValues: formattedAttachedValues,
-    from: order.from,
-    to: order.to,
-  };
-}
-
-type MappedOrder = {|
-  txId: string,
-  utxo?: string,
-  sender?: string,
-  provider?: string,
-  price: string,
-  amount: string,
-  totalValues: ?Array<FormattedTokenValue>,
-  from: any,
-  to: any,
-|};
-
-function mapOpenOrder(order: any, defaultTokenInfo: RemoteTokenInfo): MappedOrder {
-  const txId = order.utxo.split('#')[0];
-  return {
-    txId,
-    utxo: order.utxo,
-    sender: order.sender,
-    provider: order.provider,
-    ...mapOrderAssets(order, defaultTokenInfo),
-  };
-}
-
-function mapCompletedOrder(order: any, defaultTokenInfo: RemoteTokenInfo): MappedOrder {
-  return {
-    txId: order.txHash,
-    ...mapOrderAssets(order, defaultTokenInfo),
-  };
-}
-
 export default function SwapOrdersPage(props: StoresAndActionsProps): Node {
   const {
     order: orderApi,
@@ -218,11 +107,7 @@ export default function SwapOrdersPage(props: StoresAndActionsProps): Node {
     props.stores.explorers.selectedExplorer.get(network.NetworkId) ??
     fail('No explorer for wallet network');
 
-  // const openOrders = useRichOpenOrders().map(o => mapOpenOrder(o, defaultTokenInfo));
-  // const completedOrders = useRichCompletedOrders().map(o => mapCompletedOrder(o, defaultTokenInfo));
-  let { openOrders, completedOrders } = useRichOrders();
-  openOrders = openOrders.map(o => mapOpenOrder(o, defaultTokenInfo));
-  completedOrders = completedOrders.map(o => mapCompletedOrder(o, defaultTokenInfo));
+  let { openOrders, completedOrders } = useRichOrders(defaultTokenInfo);
 
   const txHashes = [...openOrders, ...completedOrders].map(o => o.txId);
   noop(props.stores.substores.ada.swapStore.fetchTransactionTimestamps({ wallet, txHashes }));
